Change result format and update chain map  in sensitivity tests.
This CL makes a minor change for the sensitivity CSV results files.
Instead of not including test results wherein the sensitivity
measurement is unreliable, the new format puts an empty string as the
sensitivity value. Further this CL adds the ability to dynamically
determine the mapping between attenuator and DUT chains eliminating the
need for hard coding the mapping in configuration files.

Test: Done
Bug: None

Change-Id: I53123440c50814366e1b7f5e1d7c4deb9a3f59fa
Signed-off-by: Omar El Ayach <oelayach@google.com>