Fix BleOpportunisticScanTest:test_max_opportunistic_scan_instances
This test is staring max_scan_instances-1 opportunistic scanners, and
one regular scanner at the end.
We had some new internal apps runing in the stack, i.e. Hearing Aid
Profile, that reduced the total number of scanners that can be started.
Reduce the max_scan_instances to accomodate for the change.

Test: BleOpportunisticScanTest:test_max_opportunistic_scan_instances
Bug: 79114848