Separate setting signal level from attachment procedure
Setting signal level should be done independently from the attachment procedure. Before this change, the signal level was set in the attach() method.

The callbox simulation is now started during the simulation class initialization. The start method in the simulation class now attaches the phone to the base station and sets up the signal level.

Bug: 134914399
Change-Id: Iaecfd959280687e3eaa0a031e3f06597c601d573