Adding BtsConfig classes to cellular simulations
The new BtsConfig classes represent a set of settings to be applied to a base station. The objective of this change if to reduce the interaction between the cellular simulation classes and the callbox controller to only one function, configure_bts. In this way, other methods just create configuration objects and pass them to configure_bts along with a base station controller handle, and the function takes care of commanding the controller to set the required configs.

Future CLs will remove configure_bts and all the methods used by it from the simulation classes into a new controller class. The ultimate objective is to make the simulation class entirely instrument agnostic.

Bug: 134914399
Change-Id: I767cb7abbaa5be691d46d51a5e0564fbb448b8b5