Creating the cellular simulator class
A cellular simulator is an abstraction layer for cellular testing equipment that allows the simulation classes to configure the testing environment. This CL defines the base class for the cellular simulator and modifies PowerCellularLabBaseTest to be able to instantiate any of the available derived classes.

The MD8475 simulator is just a wrapper to the MD8475 controller, and the cellular simulation classes just unwrap it for now. Future CLs will make the simulation interact with the cellular simulator object without using the anritsu controller contained inside.

Bug: 134914399
Change-Id: I4679b8d1e6f2b1ee1be9aa3a7f07ffe17fa14024