Move configure_bts and all its required functions to the Simulator class
This change completely removes all the Anritsu specific methods and objects from the BaseSimulation and LteSimulation classes.

The methods were moved from BaseSimulation / LteSimulation to MD8475CellularSimulator as they were. Future CLs will move parts of the methods that should be common to all the simulator implementations to the AbstractCellularSimulator class.

Bug: 134914399
Change-Id: Ied3cee353a029f0c754c5ac783c571f94f5adfc7