This CL restructures WifiRssiTest and add support for OTA RSSI tests.
This CL restructures WifiRssiTest class to be more in line with the
implementation of other Wifi performance tests. The CL also implements a
new WifiOtaRssiTest class to enable studying RSSI in varying OTA
environments. The current OTA test focus is on studying RSSI variation
at a fixed attenuation in an OTA chamber in which stirrers are
continuously moving. More tests will be added in the future.

Test: Done
Bug: None

Change-Id: I0e1388917ea91ac5d1304101a317214c1947cf6e
Signed-off-by: Omar El Ayach <oelayach@google.com>