FuchsiaDevice cleanup rebostness
This Cl addresses a few issues with Fuchsia device cleanup
if the device gets into an odd state. An easy way to repro the
current problems are to reboot the deivce in the middle of a test.

This CL adds protection in terminating SL4F and prevents dangling
threads for the syslog process.

Bug: None
Test: Run any test, reboot Fuchsia device during test
Change-Id: I55398f5bdc07b3bedca8b4e9c33c2074c03af402