Set supported MIMO modes before starting the CA simulation
To support higher MIMO modes with a single callbox, RF cards need to be shared between base stations. This can only be done if the base stations operate with the same band. In order to avoid sharing RF cards the right MIMO mode support needs to be requested. This CL sets the right support for 2 and 3 carrier tests.

Bug: 136529905
Change-Id: I59a3a0ad846d8433217a23cfc131288824934033