Adds ability to upload custom Blackbox metrics.
This CL includes and implementation of BlackboxMappedMetricLogger which
allows publishing a set of metrics, with dynamically generated names,
using the same metric logger object. Metrics can be queued up for
publishing one by one or passed to the logger in the form of a
dictionary all at once. The CL also includes an example integration of
this metric logger in WifiOTAThroughputStabilityTest.

Test: Done
Bug: None

Change-Id: I0bdf15f36c89f6c69dbe54fc9d97af0b8f88bbbd
Signed-off-by: Omar El Ayach <oelayach@google.com>