Add exception msg of TestFailure to tel_test_wrap
An issue is found that the exception message of signals.TestFailure,
including both parts of details and extras, are discarded by
tel_test_wrap. This CL intends to add the exception message back to
tel_test_wrap so that in the test summary both details and extras can be
shown correctly.

Bug: None
Test: Yes, locally
Change-Id: I5ed54458a597960244f3bb0cd37a6873bd090956