Mute test case metrics in OTA test suites
This CL mutes test case metrics in OTA test suites wherein eash test
case is defined at a single orientation. In these cases, we are only
interested in test class metrics which, for example, take the average
across all orientations. This helps keep the number of metrics uploaded
to a metric key manageable.

Test: Done
Bug: None

Change-Id: Ib58280215fc4077ed7d8ea4b5182d13dfcf1b583
Signed-off-by: Omar El Ayach <oelayach@google.com>