Augment get_func_with_retry to accept exceptions and on_retry calls.
This CL suggests an update to get_func_with_retry that adds exception
types to its arguments and includes a call to a class's on_retry
function if present. The default value for exceptions is
signals.TestFailure, maintaining the same logic as before. Similarly if
no _on_retry is defined for a class, the retry feature remains
unchanged.

Test: TBD
Bug: None

Change-Id: I8ff42a1b8f1ca1060605ed1eaa607847d6ebf430
Signed-off-by: Omar El Ayach <oelayach@google.com>