Forcibly show "VoLTE" in Settings when it's hidden
If the attribute "show_enhanced_4g_lte_mode" of the droid object is
"true" and when "hide_enhanced_4g_lte_bool = true" is found in dumpsys,
the hidden "VoLTE" (Enhanced 4G LTE Mode) option will be shown forcibly
to let scripts enable/disable VoLTE for CSFB related tests.

If "VoLTE" is hidden and "show_enhanced_4g_lte_mode" is "false" or not
available, The test cases related to CSFB will be skipped.

Bug: 138966414
Test: Yes, locally
Change-Id: Ibc4cae031a1c64dca40a892246a3e557f38f8ccc