Turn on wifi verbose by default for all tel tests
With this request to turn on verbose wifi, the simplest option was to turn wifi on for all telephony tests by default, since approximately 3/4 of our tests mention wifi.  Checked for additional logs in three cases: A test that disables wifi, A test that uses wifi, and a test that does not touch wifi. In all cases the log files were not unduly increased.

Test: Yes, locally
Bug: 142097671
Change-Id: Ibfd5586cd30119dd74899f0bc6b818d540746187