Configure special subframe config with test name parameter
The special subframe configuration needs to be a parameter for the LTE tests. The LteSimulation class will be modified to support the ssf keyword in the test names. The keyword will be followed by an underscore and the SSF config number. When the keyword is not present, the special subframe configuration will default to 6 (which is the config we have been using until now) while logging a warning message.

Bug: 143495246
Change-Id: Ibc12b1deadd28546a776d4b6d3e9e2cb9f6db8ae