Allow to configure the CFI from the test name
The CFI needs to be a parameter for the LTE tests. The LteSimulation class will be modified to support the cfi keyword in the test names. The keyword will be followed by an underscore and the CFI number. When the keyword is not present, the CFI will default to BestEffort (which is the config we have been using until now) while logging a warning message.

Bug: 143498804
Change-Id: Ia183bf80a78bd0e4ebf589a765518322352d891b