Allow to set the PHICH Resource setting from the test name
The PHICH group size needs to be a parameter for the LTE tests. The LteSimulation class will be modified to support the PHICH keyword in the test names. The keyword will be followed by an underscore and the PHICH group size. When the keyword is not present, the special subframe configuration will default to 1 (which is the config we have been using until now) while logging a warning message.

Bug: 143501301
Change-Id: I8ea76b18b28e831d29e11c3f3b1d55c9d4f291fe