Add data checking functionality to APM toggle stress test
This was accomplished by leveraging the existing airplane_mode_test
method which does in-depth checking of airplane mode toggling.  In
addition, the toggle_airplane_mode_by_adb method was always returning
true.  Updated it to have a try/catch and to check that the airplane
mode state was actually changed.

Test: Yes, locally
Bug: 143181115
Change-Id: If4226e3b33ce0b15d2086cf942043d5aa2ca651b