Fixing Build ID editing
The old code for modifying ro.build.id was making the change in
/default.prop.  It appears that the location of ro.build.id has changed
to /system/build.prop.  I updated the adb commands to modify that file
instead.  In addition, build_id_override was not handling the case of
postfix=None and recover_build_id was pulling ro.build.fingerprint first
instead of ro.vendor.build.fingerprint.  The order of checking is a
problem so all calls to recover_build_id were never actually changing
the build id back. (because getprop ro.build.id was always equal to
ro.build.fingerprint)

Test: yes, locally
Bug: 136660547
Change-Id: I850ff80b2a481206fac38a54ad420a267e6b48af