[ACTS] Add test for stations properly detecting Beacon Loss events
Turn off the AP radio to simulate loss of connectivity due to AP power
failure, reboot, etc. and verify that the STA detects the loss of the
AP.  Stopping the ap via stop_all_aps() would cause the AP to send a
Deauth message to the STAs, which is a different scenario, so use
iwconfig to turn off the radio directly.

Bug: None

Test: tested against multiple Fuchsia devices with a known
failure-to-detect-link-loss bug, and with the fix for said bug.

Change-Id: If5d7292c158ec8c23fb1194672dc8926d4cf771a