Add ability to lock an AP while tests are running.
This CL adds the ability to lock an AP while tests are running against
it. The lock is controlled by the lock_ap flag in its config. The lock
is achieved by creating a lock file in /tmp and locking it with fcntl.
Lock is released either by calling _unlock_ap explicitly or on process
termination.

Test: Done locally.
Bug: None

Change-Id: I7c7ad45c70e551056dea1d4045f1ea9059a5ec42
Signed-off-by: Omar El Ayach <oelayach@google.com>