Get the password from the code vs the security profile
Currently the password for every test is being put
in a hostapd security object and then that
password is being used directly in the tests.
However in the hostapd security object, quotes
are put around WEP ascii keys.  When that string
is then sent to fuchsia, fuchsia reads those quotes
as part of the password and fails.  So changing the
test so the password is read from the test and not
the hostapd security object.

Bug: None
Test: Ran tests and all the WEP ascii tests passed.
Change-Id: I71ba58d31edbb72122155ccf182a5077ee4eb337