Add average sensitivity metric to conducted sensitivity tests.
This CL defines a new average sensitivity metric to surface on wifi
performance dashboards. The sensitivity metrics takes the average
sensitivity of all rates of the same mode, bandwidth, number of streams,
and chain masks, and averages them. This gives an average read out on
the receiver performance of a DUT per chain and mode and is more
compact than per-rate sensitivity, i.e., more suitable for dashboards.

Test: Done
Bug: None

Change-Id: I2517ffe8f93333df7a1574ca4f990b255c3d7349
Signed-off-by: Omar El Ayach <oelayach@google.com>