[wlan][sl4f] basic SL4F test for Client Provider
Added a basic SL4F test for the Client Provider API that
simply makes each API call through the WLAN policy facade
and verifies that the commands give the appropriate error
or 'default' value (for example an empty list). The tests
will likely be removed as the implementations are completed.

Bug: 44476
Test: act.py -c <path-to-config> -tc PolicyCommandsSmokeTest

Change-Id: If22e765e29571ac18e1a0089166789e77c845a50