Fixed radvd that was creating an invalid config
There is an issue where all the varaibles were being cast
as strings and then those strings got set to None because
casting None to a string just creates a string with the
letters in it.  This made invalid settings appear in the
radvd config and caused the daemon to load.  The radvd
test was also update to reflect this change.

Bug: None
Test: Ran the test and verifed the daemon could be loaded.
Change-Id: I6d06b8d56b037835e68adfe33a2accc2519a7be1