Fix logger to no longer use the create_dir function
`utils.create_dir` only existed because ACTS ran on
python3.4, which didn't support the `exist_ok` flag,
so we had to do custom logic in `create_dir`. We've
moved on to python3.5+, so this util is no longer
necessary, and we can use the os.makedirs call
directly in those files that import this.

Bug: None
Test: Verified logger works as expected with
      python 3.6

Change-Id: I8337ddc701df9ded057cc92c01eb7b175a568ac6