Add controller for Fuchsia RegulatoryRegionService
To enable us to write region-specific WLAN tests for Fuchsia devices,
add a controller for the RegulatoryRegionService.

Design note
-----------
The RegulatoryRegionService implements two FIDL protocols. Although
we could have separate controllers for each protocol, I think we're
better off with a single controller for the whole service.

The reason I prefer a single controller is that the split in the
RRS itself is to provide granular access control. Since the ACTS
tests don't need granular access control, I think the simplicity of
a single controller wins out over the benefits of directly mirroring
the FIDL protocol structure.

Bug: None
Test: manual (see below)

Manual test
-----------
1. Flash a Fuchsia device with a custom build containing fxr/367138
   and fxr/366631
2. act.py -c ~/configs/fuchsia_testbed.config -tc InvokeSetRegionTest
3. Verify device logs show: *** QUICHE: received region #"JP"

Change-Id: Ia980dd16cdd8a7440a7d7da3e51fd989cc6adcf8