Fixed an issue when killing hostapd in access_point controller
There is an issue if wpa_supplicant fails to be killed or is
just not running, then hostapd won't be killed because the
job class throws an exception.  This would sometimes leave
the wlan interfaces up and cause issues further in the
access_point class.  This separates the two calls so if one
fails, the second will still run.

Bug: None
Test: Verfied killing a test in the middle and made sure
      hostapd was always kill even if wpa_supplicant was not
      running.
Change-Id: Icbaa159a2250f3dededd36837187f747a34e0930