Fixed a couple logic bugs on stopping the access point object
Fixed a couple issues where kills were being issued before
checking if the process is alive or not.  The is_alive
function has logic to catch the exception if a process is not
running while the kill function does not and throws an
exception if the process is not.  This fixes that case.  Also
revert the stop logic for stop the services on the AP as
without this, the processes just get respawned.

Bug: None
Test: Killed a process before stop and verified the
      exception no longer occurs.
Change-Id: I402f8ac73f5b01095faec0a75d057e26338b436e