Rename InvokeSetRegionTest
The current name of the test might be misunderstood to mean
that we're testing the functionality of the WLAN stack,
rather than just testing the facade.

Rename the test with the suffix "FacadeTest", to call out
that this is a "self-test" for the test facade, rather than
a test of end-to-end functionality.

Bug: None
Test: manual (see below)

Manual test
-----------
1. Flash a Fuchsia device with a custom build containing fxr/367138.
2. act.py -c ~/configs/fuchsia_testbed.config -tc RegulatoryRegionFacadeTest
3. Verify device logs show: *** QUICHE: received region #"JP"

Change-Id: I796df5ac2ef28ae3c6b8e9927e80a62511af6ab4