Remove dependence on golden files in WifiThroughputStabilityTest.
This CL removes the dependence of RvR golden files from the
WifiThroughputStabilityTest classes thereby avoiding the issue of stale
or missing golden files. The new version of the tests targets specific
high and low RSSI values to perform the tests at. A function that finds
the attenuation level required to achieve a target RSSI is added to wifi
performance test utils.

Test: Done
Bug: None

Signed-off-by: Omar El Ayach <oelayach@google.com>
Change-Id: I43f662ec8ca4df1cdb18e03337ee0b7362163234