Add support for Netgear RAX AP.
This CL adds support for the Netgeat RAX series of APs to wifi retail ap
class used in wifi performance tests. The RAX AP inherits most
functionality from the Netgear R7000 AP. Note that although this AP is
11ax capable, we are still referring to the bandwidth modes as VHT20,
VHT40, VHT80, and VHT160 for simplicity and compatibility with current
test code.

Test: Done
Bug: None

Signed-off-by: Omar El Ayach <oelayach@google.com>
Change-Id: Ia1b988dfdd63d1af9f2b6b3ecca488e6231d2aa3