Added timeout to base_lib to not hang on SL4F not responding
When the python requests module has a timeout of None it will
hang indefinitely.  There are many places where this can
happen but as we are seeing netstack crash this is causing any
test to stop and hang indefinitely if this is not set.  Also
added added a check to see if the device is still alive after
an SL4F call fails to return that it is offline and not just
SL4F is unreachable.

Bug: None
Test: Killed SL4F while a test was running and verified the
      test timed out appropriately and returned timed out
      when the device was still alive and device offline if
      the device was not pingable.
Change-Id: Ic9a99dbb66d1d08b27b207ecdf9617cedf4dcc74