Change the default timeout for running commands
Change the default time out for running commands on the shell
and over ssh connections.  Before this CL the default value
was 3600 seconds or 1 hour so if a simple command was run
it would wait up to an hour to timeout if the command hangs.
This can cause many tests to just hang for way too long so
this CL changes the commands to time out in 1 min.

Bug: None
Test: Ran various tests and did not notice a problem.
Change-Id: I1988af3a3d2ba62ba8717d6f60bb11588be39e9d