Changed exception to log message in Fuchsia Syslog library
The fuchsia syslog library was throwing an not needed
exception when the syslog was already started or stopped.
The library should just acknowledge the library is already
running or stopped and return.  However for good measure,
there is now a log line indicating it a double call happened
in case something breaks because of this change.

Bug: None
Test: Verified running stop or start returned and showed the
      appropriate log message.
Change-Id: I842a858a97f3c1487662ed3930623c354998a7d2